Fix errors that appeared after #2283
