+++ agent/src/main/java/com/cloud/agent/direct/download/DirectTemplateDownloaderImpl.java
