+++ agent/src/main/java/com/cloud/agent/direct/download/HttpDirectTemplateDownloader.java
