+++ api/src/main/java/org/apache/cloudstack/direct/download/DirectDownloadManager.java
