+++ core/src/main/java/com/cloud/agent/api/storage/CreateDatadiskTemplateCommand.java
