+++ core/src/main/java/com/cloud/storage/template/MetalinkTemplateDownloader.java
