+++ engine/storage/integration-test/src/test/java/org/apache/cloudstack/storage/test/DirectAgentManagerSimpleImpl.java
@@ -154,7 +154,7 @@ public class DirectAgentManagerSimpleImpl extends ManagerBase implements AgentMa
                 VmwareServerDiscoverer discover = new VmwareServerDiscoverer();
 
                 Map<? extends ServerResource, Map<String, String>> resources =
-                    discover.find(host.getDataCenterId(), host.getPodId(), host.getClusterId(), uri, userName, password, null);
+                        discover.find(host.getDataCenterId(), host.getPodId(), host.getClusterId(), uri, userName, password, null);
                 for (Map.Entry<? extends ServerResource, Map<String, String>> entry : resources.entrySet()) {
                     resource = entry.getKey();
                 }
@@ -289,5 +289,14 @@ public class DirectAgentManagerSimpleImpl extends ManagerBase implements AgentMa
 
     @Override
     public void notifyMonitorsOfRemovedHost(long hostId, long clusterId) {
+    }
+
+    @Override
+    public void disconnectWithInvestigation(long hostId, Event event) {
+
+    }
+
+    @Override
+    public void notifyMonitorsOfNewlyAddedHost(long hostId) {
     }
 }
