+++ server/src/main/java/org/apache/cloudstack/direct/download/DirectDownloadManagerImpl.java
